Multiple bug fixes

+++ src/API/PokerTime.Infrastructure/ClientEvents/Events/RoundStoryPointsSetEvent.cs
@@ -6,7 +6,7 @@ namespace PokerTime.Infrastructure.ClientEvents.Events;
 
 public class RoundStoryPointsSetEvent : IClientEvent
 {
-    public int SubmittedStoryPoints { get; set; }
+    public float SubmittedStoryPoints { get; set; }
     public int WorkItemId { get; set; }
 
     [JsonIgnore] public ClientEventType Type => ClientEventType.RoundStoryPointsSet;
